requirements.txt
aiofile
aiohttp[speedups]
appdirs
aria2p
async_generator
bs4
CairoSVG
cffi
covid
cowpy
deezloader
dnspython
emoji
Faker==8.1.2
git+https://github.com/Ncode2014/search-engine-parser
gitpython
glitch_this
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google_images_download
googletrans==3.1.0a0
gTTS
gTTS-token
hachoir
hentai
heroku3
html_telegraph_poster
httplib2
humanize

js2py
lottie
lxml
lyricsgenius
natsort
oauth2client

Pillow
psycopg2
psycopg2-binary
py7zr
pybase64
pylast
pySmartDL
python-barcode
python-dateutil
python-dotenv
pytz
qrcode
rarfile
requests
scipy
speedtest-cli
sqlalchemy<=1.3.23
telegraph
git+https://github.com/LonamiWebs/Telethon
telethon-session-sqlalchemy
telethon-tgcrypto
urbandict
validators
vcsi
wikipedia
wordcloud
youtube-dl
youtube-search
